refacto(script): rewrite `Votemarket` deployment script
-  Remove the deployment scripts from the deployment package
-  Rewrite the deployment script from scratch and store it in the votemarket package

The new version of the script is an interactive script that deploys the new votemarket protocol on all the preconfigured chains. It includes safeguards, deterministic deployment based on a pattern-salt, user-friendly logs/reminders, and all the logic included in the previous version of the script. This version is expected to be more flexible and can be run by any caller.